+++ src/main/java/com/hypherionmc/simplerpc/config/presence/PauseSection.java
@@ -1,0 +1,55 @@
+package com.hypherionmc.simplerpc.config.presence;
+
+import com.hypherionmc.simplerpc.api.rpc.RichPresenceBuilder;
+import com.hypherionmc.simplerpc.api.rpc.RichPresenceContainer;
+import com.hypherionmc.simplerpc.config.objects.RichPresenceModel;
+import dev.firstdark.rpc.enums.ActivityType;
+import shadow.hypherionmc.moonconfig.core.conversion.Path;
+import shadow.hypherionmc.moonconfig.core.conversion.SpecComment;
+import shadow.hypherionmc.moonconfig.core.fields.RandomArrayList;
+
+import java.util.ArrayList;
+
+public class PauseSection implements RichPresenceContainer {
+
+    @Path("enabled")
+    @SpecComment("Enable/Disable the Pause Event")
+    public boolean enabled = true;
+
+    @Path("presence")
+    @SpecComment("List of RPCs that will be displayed at random")
+    public RandomArrayList<RichPresenceModel> presence = new RandomArrayList<>() {{
+        add(RichPresenceModel.of(
+                ActivityType.PLAYING,
+                "In the Pause Menu",
+                "Game Paused",
+                RandomArrayList.of("mclogonew"),
+                "It's Minecraft {{game.version}}, but modded",
+                RandomArrayList.of("{{images.player}}"),
+                "{{player.name}}",
+                "https://twitch.tv/twitch",
+                new ArrayList<>()
+        ));
+    }};
+
+    @Override
+    public RichPresenceBuilder buildPresence() {
+        RichPresenceModel model = this.presence.getNextRandom().orElse(new RichPresenceModel());
+
+        return new RichPresenceBuilder()
+                .setType(model.getType())
+                .setDetails(model.getDescription())
+                .setLargeImage(model.getLargeImageKey().getNextRandom().orElse(""))
+                .setLargeImageText(model.getLargeImageText())
+                .setSmallImage(model.getSmallImageKey().getNextRandom().orElse(""))
+                .setSmallImageText(model.getSmallImageText())
+                .setState(model.getState())
+                .setButtons(model.getButtons());
+    }
+
+    @Override
+    public boolean isActive() {
+        return this.enabled;
+    }
+
+}